+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Added a method for getting the total number of seconds in a date
  - Added extensions to raise events with common event args using the data directly
  - Added property to IcdEnvironment to determine whether SSL communication is enabled
+ - Added IcdTimeZoneInfo, a very light implementation of System.TimeZoneInfo for the .NET Compact Framework
 
 ### Changed
  - Repeater changed to use configured callbacks instead of a dumb event